Restructure files
* Move main package to cmd/docker-gen
* Make everythign else part of dockergen package
