+++ utils_test.go
@@ -1,8 +1,7 @@
-package main
+package dockergen
 
 import (
 	"bytes"
-	"flag"
 	"os"
 	"strings"
 	"testing"
@@ -14,7 +13,7 @@ func TestDefaultEndpoint(t *testing.T) {
 		t.Fatalf("Unable to unset DOCKER_HOST: %s", err)
 	}
 
-	endpoint, err := getEndpoint()
+	endpoint, err := GetEndpoint("")
 	if err != nil {
 		t.Fatalf("%s", err)
 	}
@@ -29,7 +28,7 @@ func TestDockerHostEndpoint(t *testing.T) {
 		t.Fatalf("Unable to set DOCKER_HOST: %s", err)
 	}
 
-	endpoint, err := getEndpoint()
+	endpoint, err := GetEndpoint("")
 	if err != nil {
 		t.Fatal("%s", err)
 	}
@@ -41,19 +40,13 @@ func TestDockerHostEndpoint(t *testing.T) {
 
 func TestDockerFlagEndpoint(t *testing.T) {
 
-	initFlags()
 	err := os.Setenv("DOCKER_HOST", "tcp://127.0.0.1:4243")
 	if err != nil {
 		t.Fatalf("Unable to set DOCKER_HOST: %s", err)
 	}
 
 	// flag value should override DOCKER_HOST and default value
-	err = flag.Set("endpoint", "tcp://127.0.0.1:5555")
-	if err != nil {
-		t.Fatalf("Unable to set endpoint flag: %s", err)
-	}
-
-	endpoint, err := getEndpoint()
+	endpoint, err := GetEndpoint("tcp://127.0.0.1:5555")
 	if err != nil {
 		t.Fatal("%s", err)
 	}
@@ -63,8 +56,8 @@ func TestDockerFlagEndpoint(t *testing.T) {
 }
 
 func TestUnixBadFormat(t *testing.T) {
-	endpoint = "unix:/var/run/docker.sock"
-	_, err := getEndpoint()
+	endpoint := "unix:/var/run/docker.sock"
+	_, err := GetEndpoint(endpoint)
 	if err == nil {
 		t.Fatal("endpoint should have failed")
 	}
